en cours de création load more dans liste des tricks

+++ src/Controller/TrickController.php
@@ -13,13 +13,22 @@ use App\Repository\TrickRepository;
 use App\Repository\VideoRepository;
 use App\Service\MailerService;
 use App\Service\UploaderService;
+use Cocur\Slugify\Slugify;
 use Doctrine\ORM\EntityManagerInterface;
 use Knp\Component\Pager\PaginatorInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
+use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
+use Symfony\Component\Serializer\Encoder\JsonEncoder;
+use Symfony\Component\Serializer\Normalizer\AbstractNormalizer;
+use Symfony\Component\Serializer\Normalizer\ObjectNormalizer;
+use Symfony\Component\Serializer\Serializer;
+use Symfony\Component\Serializer\SerializerAwareInterface;
+use Symfony\Component\Serializer\SerializerInterface;
 
 class TrickController extends AbstractController
 {
@@ -49,7 +58,7 @@ class TrickController extends AbstractController
 
 
     #[Route(
-        path: '/trick/details/{name}',
+        path: '/trick/details/{slug}',
         name: 'app_trick_one',
         methods: ['GET', 'POST']
     )]
@@ -59,63 +68,92 @@ class TrickController extends AbstractController
         CommentaryRepository $commentaryRepository,
         VideoRepository $videoRepository,
         ImageRepository $imageRepository,
-        $name,
+        $slug,
         Request $request,
         EntityManagerInterface $entityManager
     ): Response
     {
         $trick = $tricksRepository->findOneBy(
-            criteria: ['name' => $name]
+            criteria: ['slug' => $slug]
         );
 
-        if ($trick->getId()) {
-            $commentaries = $commentaryRepository->findBy(
-                criteria: ['trick' => $trick->getId()],
-                orderBy: ['updatedAt' => 'DESC']
-            );
-            $videos = $videoRepository->findBy(
-                criteria: ['trick' => $trick->getId()]
-            );
-            $images = $imageRepository->findBy(
-                criteria: ['trick' => $trick->getId()]
-            );
+        if ($trick) {
+
+
+            if ($trick->getId()) {
+                $commentaries = $commentaryRepository->findBy(
+                    criteria: ['trick' => $trick->getId()],
+                    orderBy: ['updatedAt' => 'DESC']
+                );
+                $videos = $videoRepository->findBy(
+                    criteria: ['trick' => $trick->getId()]
+                );
+                $images = $imageRepository->findBy(
+                    criteria: ['trick' => $trick->getId()]
+                );
+            }
+
+            $commentary = new Commentary();
+            $commentaryForm = $this->createForm(CommentaryType::class, $commentary);
+            $commentaryForm->remove(name: 'createdAt');
+            $commentaryForm->remove(name: 'updatedAt');
+            $commentaryForm->remove(name: 'user');
+            $commentaryForm->remove(name: 'trick');
+
+            $commentaryForm->handleRequest($request);
+            if ($commentaryForm->isSubmitted() && $commentaryForm->isValid()) {
+                /** @var Trick $trick */
+
+                $commentary = $commentaryForm->getData();
+                $commentary->setUser($this->getUser());
+                $commentary->setTrick($trick);
+
+
+                $entityManager->persist($commentary);
+                $entityManager->flush();
+                return $this->redirectToRoute(route: 'app_trick_one', parameters: ['slug' => $trick->getSlug()]);
+            } else {
+
+                return $this->render('pages/trick/one_trick.html.twig', [
+                    'trick' => $trick,
+                    'commentaryForm' => $commentaryForm->createView(),
+                    'commentaries' => $commentaries,
+                    'videos' => $videos,
+                    'images' => $images
+                ]);
+            }
+        } else {
+
+            $this->addFlash('danger', "Cette page n'existe pas.");
+
+            return $this->redirectToRoute(route: 'app_tricks');
+
         }
-
-        $commentary = new Commentary();
-        $commentaryForm = $this->createForm(CommentaryType::class, $commentary);
-        $commentaryForm->remove(name: 'createdAt');
-        $commentaryForm->remove(name: 'updatedAt');
-        $commentaryForm->remove(name: 'user');
-        $commentaryForm->remove(name: 'trick');
-
-        $commentaryForm->handleRequest($request);
-        if ($commentaryForm->isSubmitted() && $commentaryForm->isValid()) {
-            /** @var Trick $trick */
-
-            $commentary = $commentaryForm->getData();
-            $commentary->setUser($this->getUser());
-            $commentary->setTrick($trick);
-
-
-            $entityManager->persist($commentary);
-            $entityManager->flush();
-            return $this->redirectToRoute(route: 'app_trick_one', parameters: ['name' => $trick->getName()]);
-        }
-        else {
-
-            return $this->render('pages/trick/one_trick.html.twig', [
-                'trick' => $trick,
-                'commentaryForm' => $commentaryForm->createView(),
-                'commentaries' => $commentaries,
-                'videos' => $videos,
-                'images' => $images
-            ]);
-        }
-
 
 
 
     }
+
+    #[Route(
+        path: '/tricks/load_more',
+        name: 'app_tricks_load_more',
+        methods: ['GET']
+    )]
+    public function loadMore(
+        Request $request,
+        EntityManagerInterface $entityManager,
+        SerializerInterface $serializer
+    ) : JsonResponse
+    {
+       $tricks = $entityManager->getRepository(Trick::class)->findAll();
+
+
+
+
+       return new JsonResponse(data: $serializer->serialize($tricks, 'json', ['groups' => ['trick']]), json: true);
+
+    }
+
 
     #[Route(
         path: '/trick/create',
@@ -123,7 +161,6 @@ class TrickController extends AbstractController
         methods: ['GET', 'POST']
 
 )]
-
     public function createTrick(
         Request $request,
         EntityManagerInterface $entityManager,
@@ -140,20 +177,27 @@ class TrickController extends AbstractController
         $form->remove(name: 'user');
 
         $form->handleRequest($request);
+
         if ($form->isSubmitted() && $form->isValid()) {
             /** @var Trick $trick */
 
             $trick = $form->getData();
             $trick->setUser($this->getUser());
+            $newSlug = (new Slugify())->slugify($trick->getName());
+            $trick->setSlug($newSlug);
 
-            /** @var UploadedFile $mainImageFile */
-            $mainImageFile = $form->get('mainImageFile')->getData();
-            // this condition is needed because the 'image' field is not required
-            // so the file must be processed only when a file is uploaded
-            if ($mainImageFile) {
-                $directory = $this->getParameter(name: 'images_trick_directory');
-                $trick->setImageName($uploaderService->uploadFile($mainImageFile, $directory));
+            foreach ($form->get('images') as $formChild) {
+                $uploadedImages = $formChild->get('name')->getData();
+                foreach ($uploadedImages as $uploadedImage) {
+                    $image = new Image();
+                    $image->setTrick($trick);
+                    $directory = $this->getParameter(name: 'images_trick_directory');
+                    $newFileName = $uploaderService->uploadFile($uploadedImage, $directory);
+                    $image->setName($newFileName);
+                    $entityManager->persist($image);
+                }
             }
+
 
             $entityManager->persist($trick);
             $entityManager->flush();
@@ -161,7 +205,7 @@ class TrickController extends AbstractController
             $mailMessage = $trick->getName().' '.$message;
             $mailer->sendEmail(content: $mailMessage);
             $this->addFlash('success', 'Votre trick a été ajouté');
-            return $this->redirectToRoute(route: 'app_trick_one', parameters: ['name' => $trick->getName()]);
+            return $this->redirectToRoute(route: 'app_trick_one', parameters: ['slug' => $trick->getSlug()]);
         }
         else {
 
@@ -187,61 +231,53 @@ class TrickController extends AbstractController
         UploaderService $uploaderService
     ): Response
     {
-            $trick = $trickRepository->find($id);
-            $images = $imageRepository->findBy(
-                criteria: [ 'trick' => $trick->getId()]
-            );
+        $trick = $trickRepository->find($id);
+        if ($trick) {
+            if ($trick->getUser() !== $this->getUser()) {
+                throw $this->createAccessDeniedException();
+            } else {
+                $images = $imageRepository->findBy(
+                    criteria: ['trick' => $trick->getId()]
+                );
+                $form = $this->createForm(TrickType::class, $trick);
+                $form->remove(name: 'createdAt');
+                $form->remove(name: 'updatedAt');
+                $form->remove(name: 'user');
+                $form->handleRequest($request);
 
-            $form = $this->createForm(TrickType::class, $trick);
-            $form->remove(name: 'createdAt');
-            $form->remove(name: 'updatedAt');
-            $form->remove(name: 'user');
-            $form->handleRequest($request);
 
-            if ($form->isSubmitted() && $form->isValid()) {
-                /** @var Trick $trick */
-                $trick = $form->getData();
+                if ($form->isSubmitted() && $form->isValid()) {
+                    /** @var Trick $trick */
+                    $trick = $form->getData();
+                    $newSlug = (new Slugify())->slugify($trick->getName());
+                    $trick->setSlug($newSlug);
 
-                /** @var UploadedFile $mainImageFile */
-                $mainImageFile = $form->get('mainImageFile')->getData();
-                // this condition is needed because the 'image' field is not required
-                // so the file must be processed only when a file is uploaded
-                if ($mainImageFile) {
-                    $directory = $this->getParameter(name: 'images_trick_directory');
-                    $trick->setImageName($uploaderService->uploadFile($mainImageFile, $directory));
-                }
-
-                // essai de récupération des fichiers images envoyés
-                foreach ($form->get('images') as $formChild) {
-                    $uploadedImages = $formChild->get('name')->getData();
-                    foreach ($uploadedImages as $uploadedImage) {
-                        $image = new Image();
-                        $image->setTrick($trick);
-                        $directory = $this->getParameter(name: 'images_trick_directory');
-                        $newFileName = $uploaderService->uploadFile($uploadedImage, $directory);
-                        $image->setName($newFileName);
-                        $entityManager->persist($image);
+                    foreach ($form->get('images') as $formChild) {
+                        $uploadedImages = $formChild->get('name')->getData();
+                        foreach ($uploadedImages as $uploadedImage) {
+                            $image = new Image();
+                            $image->setTrick($trick);
+                            $directory = $this->getParameter(name: 'images_trick_directory');
+                            $newFileName = $uploaderService->uploadFile($uploadedImage, $directory);
+                            $image->setName($newFileName);
+                            $entityManager->persist($image);
+                        }
                     }
-                }
-                // this condition is needed because the 'image' field is not required
-                // so the file must be processed only when a file is uploaded
-                if ($mainImageFile) {
-                    $directory = $this->getParameter(name: 'images_trick_directory');
-                    $trick->setImageName($uploaderService->uploadFile($mainImageFile, $directory));
-                }
 
-
-                $entityManager->persist($trick);
-                $entityManager->flush();
-                $this->addFlash('success', 'Votre trick a été mis à jour');
-                return $this->redirectToRoute(route: 'app_trick_one', parameters: ['name' => $trick->getName()]);
+                    $entityManager->persist($trick);
+                    $entityManager->flush();
+                    $this->addFlash('success', 'Votre trick a été mis à jour');
+                    return $this->redirectToRoute(route: 'app_trick_one', parameters: ['slug' => $trick->getSlug()]);
+                }
+                else {
+                    return $this->render('pages/trick/trick_edit.html.twig', [
+                        'form' => $form->createView(),
+                        'images' => $images,
+                        'trick' => $trick
+                    ]);
+                }
             }
-            else {
-                return $this->render('pages/trick/trick_edit.html.twig', [
-                    'form' => $form->createView(),
-                    'images' => $images
-                ]);
-            }
+        }
     }
 
 
@@ -252,6 +288,7 @@ class TrickController extends AbstractController
         methods: ['GET']
 
     )]
+    #[IsGranted('ROLE_USER')]
     public function deleteTrick(
         $id,
         TrickRepository $trickRepository,
@@ -260,8 +297,10 @@ class TrickController extends AbstractController
     {
 
         $trick = $trickRepository->find($id);
-
         if ($trick) {
+            if ($trick->getUser() !== $this->getUser()) {
+                throw $this->createAccessDeniedException();
+            }
             $entityManager->remove($trick);
             $entityManager->flush();
         }
